refactor: Artwork 패키지
