+++ fourtooncookie/src/main/java/com/startingblue/fourtooncookie/artwork/ArtworkController.java
@@ -3,7 +3,6 @@ package com.startingblue.fourtooncookie.artwork;
 import com.startingblue.fourtooncookie.artwork.dto.request.ArtworkSaveRequest;
 import com.startingblue.fourtooncookie.artwork.dto.request.ArtworkUpdateRequest;
 import com.startingblue.fourtooncookie.artwork.dto.response.ArtworkSavedResponses;
-import com.startingblue.fourtooncookie.artwork.service.ArtworkService;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
